+++ packages/sdk/src/sdk-blockchains/tezos/balance.ts
@@ -1,18 +1,23 @@
-import type { TezosProvider } from "@rarible/tezos-sdk"
+import type { TezosProvider, TezosNetwork } from "@rarible/tezos-sdk"
 import type { UnionAddress } from "@rarible/types"
 import type { AssetType } from "@rarible/api-client"
 import type { BigNumberValue } from "@rarible/utils"
 // eslint-disable-next-line camelcase
-import { get_balance } from "@rarible/tezos-sdk"
+import { get_balance, unwrap, wrap } from "@rarible/tezos-sdk"
 import BigNumber from "bignumber.js"
+import type { IBlockchainTransaction } from "@rarible/sdk-transaction"
+import { BlockchainTezosTransaction } from "@rarible/sdk-transaction"
+import type { ConvertRequest } from "../../types/balances"
 import type { MaybeProvider } from "./common"
-import { getTezosAddress, getTezosAssetType, XTZ_DECIMALS } from "./common"
+import { getRequiredProvider, getTezosAddress, getTezosAssetType } from "./common"
 
 export class TezosBalance {
 	constructor(
 		private provider: MaybeProvider<TezosProvider>,
+		private network: TezosNetwork,
 	) {
 		this.getBalance = this.getBalance.bind(this)
+		this.convert = this.convert.bind(this)
 	}
 
 	async getBalance(address: UnionAddress, assetType: AssetType): Promise<BigNumberValue> {
@@ -23,18 +28,23 @@ export class TezosBalance {
 		if (!this.provider.config.node_url) {
 			throw new Error("Node url for tezos has not been specified")
 		}
-		let balance = new BigNumber(
+		return new BigNumber(
 			await get_balance(
 				this.provider.config,
 				getTezosAddress(address),
 				tezosAssetType
 			)
 		)
+	}
 
-		if (tezosAssetType.asset_class === "XTZ") {
-			balance = balance.div(new BigNumber(10).pow(XTZ_DECIMALS))
+	async convert(request: ConvertRequest): Promise<IBlockchainTransaction> {
+		const provider = getRequiredProvider(this.provider)
+		if (request.isWrap) {
+			const tx = await wrap(provider, new BigNumber(request.value))
+			return new BlockchainTezosTransaction(tx, this.network)
+		} else {
+			const tx = await unwrap(provider, new BigNumber(request.value))
+			return new BlockchainTezosTransaction(tx, this.network)
 		}
-
-		return balance
 	}
 }