+++ packages/sdk/CHANGELOG.md
@@ -3,6 +3,50 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.45](https://github.com/rarible/sdk/compare/v0.8.44...v0.8.45) (2022-03-04)
+
+
+### Bug Fixes
+
+* polygon bid ([#155](https://github.com/rarible/sdk/issues/155)) ([1350930](https://github.com/rarible/sdk/commit/1350930f18b581401e033ad9db0db38eae317945))
+
+
+
+
+
+## [0.8.44](https://github.com/rarible/sdk/compare/v0.8.43...v0.8.44) (2022-02-25)
+
+
+### Bug Fixes
+
+* tezos balance, eth/polygon/tezos balance tests ([57c9fb7](https://github.com/rarible/sdk/commit/57c9fb70f0048bd8e22784615b30cd36a206792c))
+
+
+
+
+
+## [0.8.43](https://github.com/rarible/sdk/compare/v0.8.42...v0.8.43) (2022-02-25)
+
+
+### Bug Fixes
+
+* opensea ([8317ee6](https://github.com/rarible/sdk/commit/8317ee609c88e30d858bebaf6d214fce5b9b528c))
+
+
+
+
+
+## [0.8.42](https://github.com/rarible/sdk/compare/v0.8.41...v0.8.42) (2022-02-24)
+
+
+### Bug Fixes
+
+* eth get balance ([c6d0ed1](https://github.com/rarible/sdk/commit/c6d0ed1491483cbdb4484f768e593307e788b2ad))
+
+
+
+
+
 ## [0.8.41](https://github.com/rarible/sdk/compare/v0.8.40...v0.8.41) (2022-02-23)